backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
PyGithub==2.5.0
pydantic==2.10.3
